+++ src/importer/mutators.template.js
@@ -1,0 +1,15 @@
+import Blockly from 'blockly'
+
+/* LOCAL->> */
+const allBlockMutators = {}
+/* <<-LOCAL */
+
+for (const [blockName, mutatorObject] of Object.entries(allBlockMutators)) {
+  if(mutatorObject.helperFunction) {
+    Blockly.Extensions.registerMutator(blockName, mutatorObject, mutatorObject.helperFunction)
+  } else {
+    Blockly.Extensions.registerMutator(blockName, mutatorObject)
+  }
+}
+
+export default {}
